Squackroach (#2815)
* New Pet: The Squackroach (#2125)

<!--
This is a semi-strict format, you can add/remove sections as needed but
the order/format should be kept the same
Remove these comments before submitting
-->

---
Squackroaches are a new pet selectable in loadouts. They are the genetic
abomination-combonation of harpies and mothroaches, although they share
more in common with their harpy side. They can sing, emote in anyway a
harpy can, and can speak Valyrian Standard.

<!--
This is default collapsed, readers click to expand it and see all your
media
The PR media section can get very large at times, so this is a good way
to keep it clean
The title is written using HTML tags
The title must be within the <summary> tags or you won't see it
-->

<details><summary><h1>Media</h1></summary>
<p>

https://github.com/user-attachments/assets/c3a70e34-5ee2-4013-91a6-84d9baf027d3

</p>
</details>

---

<!--
You can add an author after the `:cl:` to change the name that appears
in the changelog (ex: `:cl: Death`)
Leaving it blank will default to your GitHub display name
This includes all available types for the changelog
-->

:cl:
- add: A new pet has been added to loadouts. Squackroaches are
gremlin-bird beasts that sing and talk to harpies!

* Squackroach Sprite Update (#2151)

<!--
This is a semi-strict format, you can add/remove sections as needed but
the order/format should be kept the same
Remove these comments before submitting
-->

# Description

---

Adds inhands, and the ability to equip to either the head or shoulder to
the Squachroach



<details><summary><h1>Media</h1></summary>
<p>

![squuack](https://github.com/user-attachments/assets/4ea07520-b9d6-43bd-8b94-edef6d7f5100)


</p>
</details>

---

# Changelog

<!--
You can add an author after the `:cl:` to change the name that appears
in the changelog (ex: `:cl: Death`)
Leaving it blank will default to your GitHub display name
This includes all available types for the changelog
-->

:cl:
- add: Squackroaches can now be worn on the head or the shoulder.
- fix: Squackroaches can now be seen being held.

* chore: Automatically update REUSE headers

* chore: Automatically update REUSE headers

* Remove language knowledge, remove stepImmune

* Removed loadout pet considering we have no loadout pets in LRP

* chore: Automatically update REUSE headers

* Move to EE folders

* chore: Automatically update REUSE headers

* chore: Automatically update REUSE headers

* correct texture path

* chore: Automatically update REUSE headers

* Remove components Goob doesn't have, account for Woundmed

* fix linter

* PR feedback

* chore: Automatically update REUSE headers

* chore: Automatically update REUSE headers

---------

Co-authored-by: dootythefrooty <137359445+dootythefrooty@users.noreply.github.com>
Co-authored-by: GoobBot <uristmchands@proton.me>